Fix offset and surface not being cleaned up properly after transforming an image...
The change at line 157 in bitmapImage is especially important, because I accidentally used topLeft() instead of mBounds.topLeft() which would otherwise result in a crash because it modifies the image with autocrop while I'm painting on it!

Also reworked the way we apply the selection in mypaint, by drawing on top of the existing surface instead of loading a new one...
this could have been a bottleneck when transforming a small section on a huge image.