Improve performance: unnecessary copy
Performance got pretty terrible on a huge surface eg. 6000x6000, this should fix that.
There's another bottleneck now though, the backup functionality. You'll notice hanging when you start a stroke, because it has to create a backup before the stroke is started.

This might not be worth looking into as it has possibly already been fixed via the undo/redo rewrite but I'll leave this note just in case.