Fix selection transforms removes previously painting content
I've reworked how we paint on top of other elements, instead of simply doing it on the main painter object. Canvaspainter will get help from a new imagecompositor class which sole purpose is to compose images and applying effects to the output.

This should prove useful when writing tests because now the canvaspainter should only care about painting while the compositor will handle how the images are layered before they're finally painted to the canvas.