QtMyPaint: make it possible to instantiate more handlers.
This makes it possible to instantiate more MPHandlers with their own lifecycle.

I've always disliked the c++ function pointer because code gets so much more verbose and just makes everything more complicated than it needs to be.  I think the parent relationship is a decent compromise as an alternative and it makes the code much easier to read and understand.

The benefit of this is also that the Handler doesn't have to be a singleton anymore, since we now have an MPHandler per instance, we can create more surfaces and be completely independent. The usecase for me is a brush preview widget, which requires its own surface, as it will otherwise paint on the canvas surface, which we don't want...