Fix camera rotation handle being offset incorrectly..
This was a bit more complicated than I first anticipated... given that there's no good way to bridge a tools UI logic and its painter, I came the conclusion that this should live independently from the painter and the tool altogether, thus I've started a UIAssists class.

I also realized that having the tool in the painter is not a good idea because the tool is not always present, nor is it as up to date as the paint operation, so you end up seeing a lot of graphical ghosting because they're out of sync.