Optimize painting by only drawing what's changed.
Basically when you fire an QWidget::update() it will update the entire canvas, this we know... what I've failed to understand is that this clears the buffer internally and means you have to paint and fill the same area in order to not see artefacts.

The trick is therefore to use the event rect, now called blitRect in CanvasPainter and only paint inside that region. In addition to that we can avoid filling the mCanvas and layer caches by following the same principle, which also improves painting performance.

There are still things to fix however... polyline is currently broken because of the above changes and the canvas cursor is causing trouble too.