Move brush presets into Preferences
Unfortunately that wasn't easy on its own, at least if we wanted to keep the same decoupling for the preference dialog.
I therefore decided to modulize the project some more in order to comply with that.

This means that we now have a new common_lib module, which consists of:
- Pencildef
- PencilError

And in the future will be able to contain more should the need arise.

The decoupling meant that I could keep using 'Status' while moving more mypaint related classes to paint_lib.

I'm also in the middle of reorganizing how we interact with the brushes.conf file and this commit is not the end of that.. it just contains enough changes to make the program continue working.