PR not to be further updated until undo/redo has been re-written.
I'm not sure I can complete this PR as the undo/redo logic is too tied to the modifications happening on the current layer.
So when we try to modify the state of another keyframe, the undo/redo state gets fucked...

I think I'll put this PR on hold, then go back working on a minimal viable undo/redo PR. We need to get that done first.