Semi Sub pixel transformation working
There's still room for improvement because when the size of transformed image changes, it does so in whole pixels, which shifts the transformed image too much. I've tried to address this by offsetting based on the shifting but it's not entirely there. You can see what I mean when you draw a 5x5 square and drag it so it crosses the edge of the pixel.

I've already spent a long time on trying to nail that problem but I'm giving it a rest for now. The overall logic and visual is much better than what we currently have.

The only downside with not being able to fix the shifting is that when you rotate the image, it still wobbles at certain angles.