Fix potential slowness when triggering hand tool
This was caused by the configurator being allocated when a tool a triggered, although it's not used for Hand tool, it's still setup.

The slowness then comes from color manager telling the mypaint preview to update on every update which causes too many updates to happen.

I'll look into throttling how often the preview is updated.