Fix stroke disappears when a selection is active
Of course with the introduction of the selection cache, I forgot about the fact that the cache needs to be updated when the frame has been modified.

We do that now by listening to the frameModified signal.

I also had to move the setModified call to after endStroke (which it probably should have been anyway), to fix the signal firing before the image had been written to.